Enable builds on p2771-0000-*
This board isn't supported by all branches, so we need to vary the
U_BOOT_BOARD axis per branch being built. Jenkins Job Builder doesn't
have a good way of doing that, so we use a Groovy script to evaluate
the list of boards to build, even though it generates a static answer
per job that's created from the template.

Signed-off-by: Stephen Warren <swarren@nvidia.com>