Always run tests even after a failed build
The build job is a matrix job, with a child job for each board. Often just
one or two builds fails and the others succeed, but in this case the build
as a whole is marked as a failure. This currently prevents the test job
from running for any board, which reduces coverage. Instead, trigger the
test job no matter what the build status is. This will ensure that
whatever tests can run actually do. This will cause a bit more noise as
test jobs fail because the builds they attempt to pick up artifacts from
did not succeed, but that's not hard to deal with given curernt volumes.

Signed-off-by: Stephen Warren <swarren@nvidia.com>